Sửa long lại thành int 	modified:   FloginFE_BE/backend/src/main/java/com/flogin/controller/ProductController.java 	modified:   FloginFE_BE/backend/src/main/java/com/flogin/dto/ProductDTO.java 	modified:   FloginFE_BE/backend/src/main/java/com/flogin/model/Product.java 	modified:   FloginFE_BE/backend/src/main/java/com/flogin/repository/ProductRepository.java 	modified:   FloginFE_BE/backend/src/main/java/com/flogin/service/ProductService.java 	modified:   FloginFE_BE/backend/src/main/java/com/flogin/service/impl/ProductServiceImpl.java 	modified:   FloginFE_BE/backend/src/test/java/com/flogin/service/ProductServiceMockTest.java

+++ FloginFE_BE/backend/src/main/java/com/flogin/controller/ProductController.java
@@ -70,7 +70,7 @@ public class ProductController {
      * @throws ResourceNotFoundException nếu không tìm thấy product
      */
     @GetMapping("/{id}")
-    public ResponseEntity<ProductDTO> getProductById(@PathVariable Long id) {
+    public ResponseEntity<ProductDTO> getProductById(@PathVariable int id) {
         ProductDTO product = productService.getProductById(id);
         return ResponseEntity.ok(product);
     }
@@ -86,7 +86,7 @@ public class ProductController {
      */
     @PutMapping("/{id}")
     public ResponseEntity<ProductDTO> updateProduct(
-            @PathVariable Long id,
+            @PathVariable int id,
             @Valid @RequestBody ProductDTO dto) {
         ProductDTO updated = productService.updateProduct(id, dto);
         return ResponseEntity.ok(updated);
@@ -101,7 +101,7 @@ public class ProductController {
      * @throws ResourceNotFoundException nếu không tìm thấy product
      */
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteProduct(@PathVariable int id) {
         productService.deleteProduct(id);
         return ResponseEntity.noContent().build();
     }
