+++ FloginFE_BE/backend/src/main/java/com/flogin/service/ProductService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface ProductService {
     ProductDTO createProduct(ProductDTO dto);
     List<ProductDTO> getAllProducts();
-    ProductDTO getProductById(Long id);
-    ProductDTO updateProduct(Long id, ProductDTO dto);
-    void deleteProduct(Long id);
+    ProductDTO getProductById(int id);
+    ProductDTO updateProduct(int id, ProductDTO dto);
+    void deleteProduct(int id);
 }
